Get rid of TextView+Keyboard.swift
