+++ Watt/TextView+Input.swift
@@ -7,6 +7,19 @@
 
 import Cocoa
 
+extension TextView {
+    override func keyDown(with event: NSEvent) {
+        NSCursor.setHiddenUntilMouseMoves(true)
+
+        if inputContext?.handleEvent(event) ?? false {
+            return
+        }
+
+        // Don't know if handleEvent ever returns false here. Just want to know about it.
+        fatalError("keyDown: inputContext didn't handle this event: \(event)")
+    }
+}
+
 extension TextView: NSTextInputClient {
     func insertText(_ string: Any, replacementRange: NSRange) {
         guard let range = getReplacementRange(for: replacementRange) else {
